checkers/example-service/requirements.txt
anyio==4.6.0
cachetools==5.5.0
certifi==2024.8.30
google-auth==2.35.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
rsa==4.9
sniffio==1.3.1
